separate float from coords
