+++ qubed/src/coordinates/mod.rs
@@ -1,8 +1,10 @@
+pub mod floats;
 pub mod integers;
 pub mod ops;
 pub mod strings;
 use std::hash::Hash;
 
+use floats::FloatCoordinates;
 use integers::IntegerCoordinates;
 use strings::StringCoordinates;
 
@@ -23,10 +25,10 @@ pub enum Coordinates {
     Mixed(Box<MixedCoordinates>),
 }
 
-#[derive(Debug, Clone, PartialEq)]
-pub enum FloatCoordinates {
-    List(TinyVec<f64, 4>),
-}
+// #[derive(Debug, Clone, PartialEq)]
+// pub enum FloatCoordinates {
+//     List(TinyVec<f64, 4>),
+// }
 
 pub enum CoordinateTypes {
     Integer(i32),
@@ -266,52 +268,52 @@ impl Default for Coordinates {
     }
 }
 
-impl FloatCoordinates {
-    fn extend(&mut self, _new_coords: &FloatCoordinates) {
-        todo!()
-    }
-    fn append(&mut self, _new_coord: f64) {
-        todo!()
-    }
+// impl FloatCoordinates {
+//     fn extend(&mut self, _new_coords: &FloatCoordinates) {
+//         todo!()
+//     }
+//     fn append(&mut self, _new_coord: f64) {
+//         todo!()
+//     }
 
-    fn len(&self) -> usize {
-        match self {
-            FloatCoordinates::List(list) => list.len(),
-        }
-    }
-    pub(crate) fn to_string(&self) -> String {
-        match self {
-            FloatCoordinates::List(list) => {
-                list.iter().map(|v| v.to_string()).collect::<Vec<String>>().join("/")
-            }
-        }
-    }
+//     fn len(&self) -> usize {
+//         match self {
+//             FloatCoordinates::List(list) => list.len(),
+//         }
+//     }
+//     pub(crate) fn to_string(&self) -> String {
+//         match self {
+//             FloatCoordinates::List(list) => {
+//                 list.iter().map(|v| v.to_string()).collect::<Vec<String>>().join("/")
+//             }
+//         }
+//     }
 
-    pub(crate) fn hash(&self, hasher: &mut std::collections::hash_map::DefaultHasher) {
-        "floats".hash(hasher);
-        match self {
-            FloatCoordinates::List(list) => {
-                for val in list.iter() {
-                    val.to_bits().hash(hasher);
-                }
-            }
-        }
-    }
-}
+//     pub(crate) fn hash(&self, hasher: &mut std::collections::hash_map::DefaultHasher) {
+//         "floats".hash(hasher);
+//         match self {
+//             FloatCoordinates::List(list) => {
+//                 for val in list.iter() {
+//                     val.to_bits().hash(hasher);
+//                 }
+//             }
+//         }
+//     }
+// }
 
-impl Default for FloatCoordinates {
-    fn default() -> Self {
-        FloatCoordinates::List(TinyVec::new())
-    }
-}
+// impl Default for FloatCoordinates {
+//     fn default() -> Self {
+//         FloatCoordinates::List(TinyVec::new())
+//     }
+// }
 
-impl From<f64> for Coordinates {
-    fn from(value: f64) -> Self {
-        let mut vec = TinyVec::new();
-        vec.push(value);
-        Coordinates::Floats(FloatCoordinates::List(vec))
-    }
-}
+// impl From<f64> for Coordinates {
+//     fn from(value: f64) -> Self {
+//         let mut vec = TinyVec::new();
+//         vec.push(value);
+//         Coordinates::Floats(FloatCoordinates::List(vec))
+//     }
+// }
 
 // ------------- Intersection ------------------
